6. Review: thu gon code o menu (tat ca cac view) + sua logic (Product, Customer - check loi khi xoa khoa ngoai,)

+++ module2/src/ra/phone_store_manager/business/impl/ProductServiceImpl.java
@@ -40,6 +40,13 @@ public class ProductServiceImpl implements IProductService {
             return false;
         }
 
+        if (productDAO.isProductInInvoice(id)) {
+            System.out.println(Color.DO+
+                    "* Lỗi: Sản phẩm đã tồn tại trong lịch sử giao dịch không thể xóa!"
+                    + Color.RESET);
+            return false;
+        }
+
         return productDAO.deleteProduct(id);
     }
